requirements.txt
av==9.0.1
ConfigArgParse==1.5.3
einops==0.4.1
ffmpeg==1.4
ffmpeg_python==0.2.0
h5py==3.7.0
ipdb==0.13.9
librosa==0.9.1
loguru==0.6.0
matplotlib==3.5.1
numpy==1.21.5
pandas==1.3.5
Pillow==9.2.0
scikit_learn==1.1.2
scipy==1.7.3
six==1.16.0
torch==1.9.0+cu111
torchaudio==0.9.0
torchvision==0.10.0+cu111
tqdm==4.63.0
